Added in user interaction with simple arithmetic calculations.
With the functions addition, subtraction, multiplication, and division, I add a user interface for the simple calculator where the user is prompted to enter two integer values followed by another prompt to enter a operator. The sum is then given to the user and they are asked if they want to continue or quit the calculator.